[WIP]: big commit feat: remove register component and related translations
feat: update arrow service to include pagination and post functionality

feat: enhance user service with get and update user methods

feat: add French translations for account creation and login

feat: implement JWT authentication endpoints in the API

feat: create modal component for displaying messages and errors

feat: add auth interceptor to manage JWT tokens in requests

feat: implement token management service for authentication

feat: create account component for user account management

feat: implement login component for user authentication

feat: create search component for finding arrows with pagination

feat: add functionality to add new arrows to the database

+++ api/endpoints/arrows.py
@@ -1,16 +1,18 @@
-from flask import request
+from flask import jsonify, make_response, request
 from marshmallow import ValidationError
-from werkzeug.exceptions import BadRequest
+from werkzeug.exceptions import BadRequest, Forbidden
 
 from app import db
 from models.arrow import Arrow, arrow_schema, arrows_schema
+from models.arrow import arrow_public_schema, arrows_public_schema
 from utils import Utils
 
 
-# @require_auth
+@Utils.require_auth
 def search(offset, limit, filters):
     try:
         query = Utils.build_query_filters(Arrow, filters)
+        total = Arrow.query.filter(*query).count()
         arrows = Arrow.query.filter(*query)\
             .limit(limit).offset(offset).all()
     except AttributeError:
@@ -19,10 +21,19 @@ def search(offset, limit, filters):
 
     if not arrows:  # pragma: no cover
         return [], 204
-    else:
-        return arrows_schema.dump(arrows)
+
+    schema = arrows_schema if Utils.is_admin() else arrows_public_schema
+    serialized = schema.dump(arrows)
+
+    response = make_response(jsonify(serialized))
+    response.headers["X-Total-Count"] = total
+    response.headers["X-Limit"] = limit
+    response.headers["X-Offset"] = offset
+
+    return response
 
 
+@Utils.require_auth
 def post(arrow_data=None, **kwargs):
     if arrow_data is None:
         arrow_data = request.get_json()
@@ -30,6 +41,8 @@ def post(arrow_data=None, **kwargs):
         data = arrow_schema.load(arrow_data)
     except ValidationError as err:
         raise BadRequest(description=str(err))
+    if not hasattr(data, 'user_id'):
+        data['user_id'] = Utils.get_userid()
     arrow = Arrow(**data)
     db.session.add(arrow)
     db.session.commit()
@@ -39,18 +52,23 @@ def post(arrow_data=None, **kwargs):
     }
 
 
-# @require_auth
+@Utils.require_auth
 def get(arrow_id):
-    arrow_in_DB = Utils.getObjectInDB(Arrow, arrow_id)
-    arrow = arrow_schema.dump(arrow_in_DB)
+    arrow_in_DB = Utils.get_from_db(Arrow, arrow_id)
+    if Utils.is_admin():
+        arrow = arrow_schema.dump(arrow_in_DB)
+    else:
+        arrow = arrow_public_schema.dump(arrow_in_DB)
     return arrow, 200, {
         'Location': f'{request.base_url}/arrows/{arrow_in_DB.id}',
     }
 
 
-# @require_auth
+@Utils.require_auth
 def put(arrow_id, **kwargs):
-    arrow_in_DB = Utils.getObjectInDB(Arrow, arrow_id)
+    arrow_in_DB = Utils.get_from_db(Arrow, arrow_id)
+    if not Utils.is_admin() and arrow_in_DB.user_id != Utils.get_userid():
+        raise Forbidden(description='You can only edit your own arrows.')
     arrow_data = kwargs.get('body', {})
     try:
         data = arrow_schema.load(arrow_data)
@@ -65,9 +83,11 @@ def put(arrow_id, **kwargs):
     }
 
 
-# @require_auth
+@Utils.require_auth
 def delete(arrow_id):
-    arrow = Utils.getObjectInDB(Arrow, arrow_id)
+    arrow = Utils.get_from_db(Arrow, arrow_id)
+    if not Utils.is_admin() and arrow.user_id != Utils.get_userid():
+        raise Forbidden(description='You can only delete your own arrows.')
     db.session.delete(arrow)
     db.session.commit()
     return None, 204
